add seance model and service and controller

+++ src/schoolmanager/BackEnd/Model/Group.java
@@ -26,7 +26,7 @@ public class Group {
     public Group() {
     }
 
-    public Group(int id) {
+    public Group(Long id) {
         this.id = id;
     }
